Correcting some easy-to-fix warnings

+++ src/graphics/MediaViewStateMachine.cpp
@@ -17,7 +17,7 @@ const static float PI = 3.14159265f;
 
 const double startAngle = 3 * M_PI/4.0;
 const double endAngle = startAngle + 6 * (M_PI/4.0);
-const double numItems = 3;
+const size_t numItems = 3;
 const Color bgColor(0.4f, 0.425f, 0.45f, 0.75f);
 const Color fillColor(0.4f, 0.8f, 0.4f, 0.7f);
 const Color handleColor(0.65f, 0.675f, 0.7f, 1.0f);
@@ -138,9 +138,9 @@ void MediaViewStateMachine::AutoFilter(OSCState appState, const HandLocation& ha
       
       // VOLUME UPDATE
       //m_mediaViewEventListener(&MediaViewEventListener::OnUserChangedVolume)(calculateVolumeDelta(dHandRoll.dTheta));
-      const float DEADZONE = 0.4;
-      const float MAX = 1.0;
-      const float MAX_VELOCTY = 0.7;
+      const float DEADZONE = 0.4f;
+      const float MAX = 1.0f;
+      const float MAX_VELOCTY = 0.7f;
       
       if( !m_hasRoll ) {
         m_startRoll = dHandRoll.absoluteRoll;
